requirements.txt
click==7.1.2
label-studio-ml==1.0.8
redis
rq==1.10.1
gunicorn==20.1.0
werkzeug==2.0.2
itsdangerous==2.0.1

label-studio-converter
sentencepiece
boto3
openmim
opencv-contrib-python-headless
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.8.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.8.1
# tqdm>=4.64.0